separate CSS file in the styles folder
